# api docs/swagger.yaml
openapi: 3.0.0
info:
  title: Video share API
  version: 1.0.0
  description: Applicaion for sharing videos
...
servers:
  - url: 'http://localhost:8080/api/v1'
  - url: 'http://192.168.0.26:8080/api/v1'
components:
  securitySchemes:
    token:
      type: apiKey
      in: cookie
      name: token 
    refreshToken:
      type: apiKey
      in: cookie
      name: refreshToken  
...
  requestBodies:
...
    AuthRegisterBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
                format: password
              email:
                type: string
                format: email
...
    AuthLoginBody:
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              username:
                type: string
              password:
                type: string
                format: password
...
security:
  - token: []
  - refreshToken: []
paths:
...
  /register:
    post:
      summary: Register a new user
      tags:
        - Auth
      security: []
      requestBody:
        $ref: '#/components/requestBodies/AuthRegisterBody'
      responses:
        '200':
          description: Successful operation
        '500':
          description: Internal server error
  /login:
    post:
      summary: Login user
      tags:
        - Auth
      security: []
      requestBody:
        $ref: '#/components/requestBodies/AuthLoginBody'
      responses:
        '200':
          description: Successful operation
        '400':
          description: Wrong password
        '404':
          description: User not found
  /logout:
    patch:
      summary: Logout user
      tags:
        - Auth
      responses:
        '200':
          description: Successful operation
        '404':
          description: User not found
...